+++ examples/index.tsx
@@ -36,6 +36,7 @@ import Segmented from './LineChart/Segmented';
 import SegmentedDataSetCurved from './LineChart/SegmentedDataSetCurved';
 import SegmentedDataSetChart from './LineChart/SegmentedDataSet';
 import HighlightedRange from './LineChart/HighlightedRange';
+import GradientLineAndLabel from './LineChart/GradientLineAndLabel';
 
 const Examples = () => {
   const [selectedFooterButton, setSelectedFooterButton] = useState(0);
@@ -105,6 +106,9 @@ const Examples = () => {
         <ChartWithPointer />
         <Separator />
 
+        <GradientLineAndLabel />
+        <Separator />
+
         <AreaChartDynamicData />
         <Separator />
 
@@ -135,7 +139,7 @@ const Examples = () => {
 
         <SegmentedDataSetCurved />
         <Separator />
-        
+
         <HighlightedRange />
         <Separator />
         <Separator />